+++ actionpack/lib/action_view/helpers/form_tag_helper.rb
@@ -564,6 +564,14 @@ module ActionView
         text_field_tag(name, value, options.stringify_keys.update("type" => "date"))
       end
 
+      # Creates a text field of type "month".
+      #
+      # ==== Options
+      # * Accepts the same options as text_field_tag.
+      def month_field_tag(name, value = nil, options = {})
+        text_field_tag(name, value, options.stringify_keys.update("type" => "month"))
+      end
+
       # Creates a text field of type "url".
       #
       # ==== Options